Added some replication functionality and fixed bugs
Fixed an issue with report id not corresponding to index in reports slice
Initialized Reading Variable in replication slice
Added an Add Reading button to test that replication works
Fixed an issue with ALL of our selectors. We must access state using the keywords in store. More about this later.