Attempted change to End Protocol Button
This is one of the ways I thought to do it initially. Other methods involved using useState and trying to update the button in that way, but that just lead to many more errors, and the button still would never rerender. My thinking here goes that if the button re-renders, it will call allValid() again, and update. However, I can't figure out how to get the button to re-render when the StandardReplicationRow calls onChange.